Conflicts:
	src/main/resources/com/clevel/selos/messages_en_US.properties

+++ src/main/java/com/clevel/selos/businesscontrol/CustomerInfoControl.java
@@ -85,6 +85,35 @@ public class CustomerInfoControl extends BusinessControl {
                     cV.setPercentShareSummary(BigDecimal.ZERO);
                 }
             }
+
+            //for show jurLv
+            if(cV.getIsCommittee() == 1){
+                for(CustomerInfoView cusView : customerInfoViewList){
+                    if(cusView.getId() == cV.getCommitteeId()){
+                        cV.setJurLv(cV.getReference().getDescription()+" of "+cusView.getFirstNameTh()+" "+cusView.getLastNameTh());
+                    }
+                }
+            } else {
+                cV.setJurLv("-");
+            }
+
+            //for show indLv
+            if(cV.getIsSpouse() == 1){ // is spouse
+                for(CustomerInfoView cusView : customerInfoViewList){ // is main spouse
+                    if(cusView.getSpouseId() == cV.getId()){
+                        cV.setIndLv(cV.getReference().getDescription()+" of "+cusView.getFirstNameTh()+" "+cusView.getLastNameTh());
+                        if(cusView.getIsCommittee() == 1){ // is main spouse is committee
+                            for(CustomerInfoView cusViewJur : customerInfoViewList){
+                                if(cusViewJur.getId() == cusView.getCommitteeId()){
+                                    cV.setJurLv(cusView.getReference().getDescription()+" of "+cusViewJur.getFirstNameTh()+" "+cusViewJur.getLastNameTh());
+                                }
+                            }
+                        }
+                    }
+                }
+            } else {
+                cV.setIndLv("-");
+            }
         }
 
         List<CustomerInfoView> borrowerCustomerList = customerTransform.transformToBorrowerViewList(customerInfoViewList);
@@ -291,7 +320,9 @@ public class CustomerInfoControl extends BusinessControl {
 
         //for check customer ncb
         NCB ncb = ncbDAO.findNcbByCustomer(customer.getId());
-        ncbDAO.delete(ncb);
+        if(ncb != null){
+            ncbDAO.delete(ncb);
+        }
 
         customerDAO.delete(customer);
     }
